Project started and index mostly finished.
I have set the coding standards in my <head> section and have added the headings as well as the paragraphs that come along with them with all the font types the require. I lastly added the footer and then started on my dr_info.html page, which is where I left off.